# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Roadmap for Jira API",
        "description": "Endpoint mapping for Roadmap for Jira App.",
        "contact": {
            "email": "contact@mrrafael.ca"
        },
        "version": "2.0"
    },
    "servers": [
        {
            "url": "http://localhost:3030",
            "description": "Server"
        }
    ],
    "paths": {
...
        "/api/v1/roadmap/share": {
            "put": {
                "tags": [
                    "Roadmap"
                ],
                "description": "Share the current roadmap with a new user by email.",
                "operationId": "Share",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddShareParam"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object with all shared users",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ShareModelList"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object containing the error message",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorModel"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Roadmap"
                ],
                "description": "Remove the shared user from current roadmap by user email.",
                "operationId": "RemoveShare",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RemoveShareParam"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object with all shared users",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ShareModelList"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object containing the error message",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorModel"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AddShareParam": {
                "required": [
                    "guid",
                    "email",
                    "readOnly"
                ],
                "properties": {
                    "guid": {
                        "type": "string",
                        "example": "455bbfd4-abcd-1234-8345-2d29d164699b"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@test.com"
                    },
                    "readOnly": {
                        "type": "boolean",
                        "example": "true"
                    }
                },
                "type": "object"
            },
            "RemoveShareParam": {
                "required": [
                    "guid",
                    "email"
                ],
                "properties": {
                    "guid": {
                        "type": "string",
                        "example": "455bbfd4-abcd-1234-8345-2d29d164699b"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@test.com"
                    }
                },
                "type": "object"
            },
            "ShareModel": {
                "required": [
                    "id",
                    "read_only",
                    "user_email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "3"
                    },
                    "read_only": {
                        "type": "integer",
                        "example": "0"
                    },
                    "user_email": {
                        "type": "string",
                        "example": "john@test.com"
                    }
                },
                "type": "object"
            },
            "ShareModelList": {
                "required": [
                    "list"
                ],
                "properties": {
                    "list": {
                        "type": "array",
                        "items": {
                            "oneOf": [
                                {
                                    "$ref": "#/components/schemas/ShareModel"
                                }
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
            "ErrorModel": {
                "required": [
                    "code",
                    "error"
                ],
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "42S22"
                    },
                    "error": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "message": {
                        "type": "string",
                        "example": "Complementary Error message"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}